enh: move signal extraction functions to a better location
